+++ zorka-agent/src/main/java/com/jitlogic/zorka/agent/AgentInstance.java
@@ -18,6 +18,8 @@ package com.jitlogic.zorka.agent;
 
 import com.jitlogic.zorka.integ.nagios.NagiosAgent;
 import com.jitlogic.zorka.integ.nagios.NagiosLib;
+import com.jitlogic.zorka.integ.snmp.SnmpLib;
+import com.jitlogic.zorka.integ.syslog.SyslogLib;
 import com.jitlogic.zorka.spy.MainSubmitter;
 import com.jitlogic.zorka.spy.SpyInstance;
 import com.jitlogic.zorka.spy.SpyLib;
@@ -83,6 +85,9 @@ public class AgentInstance {
     private SpyLib spyLib = null;
     private SpyInstance spyInstance = null;
 
+    private SyslogLib syslogLib = null;
+    private SnmpLib snmpLib = null;
+
     private Properties props;
 
 
@@ -132,6 +137,18 @@ public class AgentInstance {
             log.info("Zorka SPY is diabled. No loaded classes will be transformed in any way.");
         }
 
+        if ("yes".equalsIgnoreCase(props.getProperty(SYSLOG_ENABLE))) {
+            log.info("Enabling Syslog subsystem ....");
+            syslogLib = new SyslogLib();
+            zorkaAgent.installModule("syslog", syslogLib);
+        }
+
+        if ("yes".equalsIgnoreCase(props.getProperty(SNMP_ENABLE))) {
+            log.info("Enabling SNMP subsystem ...");
+            snmpLib = new SnmpLib();
+            zorkaAgent.installModule("snmp", snmpLib);
+        }
+
         zorkaAgent.loadScriptDir(props.getProperty(ZORKA_CONF_DIR), ".*\\.bsh$");
 
         if ("yes".equalsIgnoreCase(props.getProperty(ZABBIX_ENABLE))) {
@@ -159,6 +176,14 @@ public class AgentInstance {
 
     public ZorkaBshAgent getZorkaAgent() {
         return zorkaAgent;
+    }
+
+    public SyslogLib getSyslogLib() {
+        return syslogLib;
+    }
+
+    public SnmpLib getSnmpLib() {
+        return snmpLib;
     }
 
 }