+++ zorka-agent/src/main/java/com/jitlogic/zorka/agent/ZorkaConfig.java
@@ -54,6 +54,10 @@ public class ZorkaConfig {
     public final static String SPY_ENABLE = "spy";
     public final static String SPY_DEBUG = "spy.debug";
 
+    public final static String SYSLOG_ENABLE = "syslog";
+
+    public final static String SNMP_ENABLE = "snmp";
+
     public final static String ZABBIX_ENABLE = "zabbix";
 
     public final static String NAGIOS_ENABLE = "nagios";