+++ CHANGES.md
@@ -1,6 +1,13 @@
 
+Zorka 0.5 (2012-12-07)
+----------------------
+
+ * syslog protocol support;
+ * snmp protocol support (traps and value get interface);
+
+
 Zorka 0.4 (2012-11-26)
----------
+----------------------
 
  * documentation updates, cleanups and fixes (as usual);
  * nagios NRPE protocol support;
@@ -9,7 +16,7 @@ Zorka 0.4 (2012-11-26)
 
 
 Zorka 0.3 (2012-11-10)
---------------------------------
+----------------------
 
  * basic ranking framework (working with Zorka MethodCallStats);
  * implement missing collectors of ZorkaSpy;
@@ -29,7 +36,7 @@ Zorka 0.2 (2012-11-04)
 
 
 Zorka 0.1 (2012-09-19)
----------
+----------------------
 
  * initial release;
  * zabbix integration;